Refactor: split logic: vote checking is done by Engine, append-entries is done by FollowingHandler
